data_pipeline/inventory_control_cleaning_function_source/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
numpy
openpyxl
pyarrow